feat: expose a dedicated 'buffers' user-mode
There's a slight breaking change with this approach:

Let's say you have the `b` key to trigger the buffers mode.
`b3` was used to go the third buffer. Now it's the other way around,
`3b`. It uses native `count` so it follows the same logic than `3g` to
go to the third line. It's an improvement since you're not limited to
10 buffers anymore.